[WebXR][GBM] GBM texture export only allows 2 formats https://bugs.webkit.org/show_bug.cgi?id=302275
Reviewed by Adrian Perez de Castro.

We should not constraint ourselves to use DRM_FORMAT_ARGB8888 and/or
DRM_FORMAT_XRGB8888 when using GBM to export textures as the
GraphicsContext is able to import many more. Different hardware and/or
drivers might not be able to support any of the current ones.

Added a couple more of formats for both alpa and non-alpha use cases. We
select them based on the order we specify and the list of supported
formats coming from the platform display.

No need for new tests as this does not change functionality, it extends
support for additional hardware.

* Source/WebKit/UIProcess/XR/openxr/OpenXRLayer.cpp:
(WebKit::OpenXRLayer::exportOpenXRTextureGBM):

Canonical link: https://commits.webkit.org/302838@main