[GFC] Move GridFormattingContextIntegrationEnabled to stable https://bugs.webkit.org/show_bug.cgi?id=302264 rdar://164406235
Reviewed by Alan Baradlay.

The GFC codepath now has an implementation for some type of grid
content, so we can move the feature flag to stable. This will allow the
eligbility code to determine whether or not we can take GFC to run. For
the the vast majority of grid content we will not take GFC, but as we
build up our implementation we can start removing more of these
restrictions one by one.

Canonical link: https://commits.webkit.org/302846@main