ASAN_ILL | WebCore::ImmutableStyleProperties::createDeduplicating; WebCore::createStyleProperties; WebCore::CSSParser::parseInlineStyleDeclaration https://bugs.webkit.org/show_bug.cgi?id=295927 rdar://155792201
Reviewed by Simon Fraser.

AlreadyHashed::avoidDeletedValue() always returns the passed hash, as
the or-bitwise operator doesn't really do anything. If the expected
behavior of this method is to avoid the special 0xFFFFFFFF value,
then I strongly suspect that what was intended was to use a xor-bitwise
operation; this will indeed flip the topmost bit if the hash is
0xFFFFFFFF but leave it unchanged for any other value (except 0,
of course).

ImmutableStyleProperties should use this method when adding elements
to its hash map, to make sure that it doesn't accidentally try to
insert a value whose hash key is 0xFFFFFFFF. Otherwise it will
hit an assertion.

* LayoutTests/fast/css/immutable-properties-crash-expected.txt: Added.
* LayoutTests/fast/css/immutable-properties-crash.html: Added.
* Source/WTF/wtf/text/StringHash.h:
(WTF::AlreadyHashed::avoidDeletedValue):
* Source/WebCore/css/ImmutableStyleProperties.cpp:
(WebCore::ImmutableStyleProperties::createDeduplicating):

Originally-landed-as: 297297.3@webkit-2025.7-embargoed (c54787bc3732). rdar://164279182
Canonical link: https://commits.webkit.org/302886@main