Retain the TypeDefinition underlying the Wasm::Type in Wasm::Global and Wasm::Table https://bugs.webkit.org/show_bug.cgi?id=297958 rdar://159278266
Reviewed by Yusuke Suzuki.

Wasm::Global and Wasm::Table both have a field containing a Wasm::Type describing the
value of the global or the contents of the table. Wasm::Type has a field `index` of type
`TypeIndex`. Formally, `TypeIndex` is an alias for `uintptr_t`. However, the field may
actually be a pointer to the `TypeDefinition` instance corresponding to the type.

Because TypeDefinitions are refcounted but Wasm::Type contains a raw unmanaged pointer, it
is possible for a pointer in Wasm::Type to outlive the TypeDefinition it points to. This
can happen with Wasm Globals and Tables because an exported global or a table (if it's
empty) may outlive their associated Wasm instance when they are retained by JavaScript
code. In that situation TypeDefinitions referenced by the instance may be
garbage-collected while pointers to them from the globals or tables still exist. After
that, an attempt to set the value of the global or a table entry will dereference the
dangling pointer as the value is validated against the type of the global or the table.

This patch adds an explicit counted TypeDefinition reference to Wasm::Global
and Wasm::Table to ensure the TypeDefinition lives long enough.

Ideally, we might want to redesign Wasm::Type to have it carry a proper counted reference
to its TypeDefinition instead of a disguised raw pointer. But that is a larger
undertaking, while in the meantime this patch fixes the known vulnerability.

Originally-landed-as: 297297.393@safari-7622-branch (b3e1519a9399). rdar://164277125
Canonical link: https://commits.webkit.org/302928@main