http/tests/media/fairplay/legacy-fairplay-mse-v2.html is an intermittent failure https://bugs.webkit.org/show_bug.cgi?id=302425 rdar://164587106
Reviewed by Youenn Fablet.

The SourceBuffer's appendBuffer operation is asynchronous, there's no guarantee
that the EME/webkitkey event will fire in a particular order in regards to
`updateend` ; so we wait for updateend to be fired first before completing
the key negotiations. This guarantee the text output order.

* LayoutTests/http/tests/media/fairplay/legacy-fairplay-mse-v2-expected.txt:
* LayoutTests/http/tests/media/fairplay/legacy-fairplay-mse-v2.html:
* LayoutTests/http/tests/media/fairplay/legacy-fairplay-mse-v3-expected.txt:
* LayoutTests/http/tests/media/fairplay/legacy-fairplay-mse-v3.html:

Canonical link: https://commits.webkit.org/302972@main