[webkitpy] Executive.run_command should log command before running. https://bugs.webkit.org/show_bug.cgi?id=302485 rdar://164649390
Reviewed by Jonathan Bedard.

Executive.run_command only logs the command being run after it completes. When
debugging infrastructure issues, it helps to know what command was running and
when.

So, this changes the logging behavior of Executive.run_command such that it
logs what command is about to be run before it runs.

* Tools/Scripts/webkitpy/common/system/executive.py:
    (Executive.run_command): Change logging behavior.

Canonical link: https://commits.webkit.org/303002@main