[JSC] Cleaning up iterator related code https://bugs.webkit.org/show_bug.cgi?id=302562 rdar://164768602
Reviewed by Yijia Huang.

This is fixing a bug introduced in 302914@main.
JSIteratorHelperPrototype is using generator too, so it needs to be
changed too. This is caught by test262.
We also change iterator result object inlineCapacity to align it to the
JS generated ones.

* Source/JavaScriptCore/builtins/JSIteratorHelperPrototype.js:
(next):
(return):
* Source/JavaScriptCore/runtime/IteratorOperations.cpp:
(JSC::createIteratorResultObjectStructure):

Canonical link: https://commits.webkit.org/303065@main