[MathML] Respect inline-start for mpadded lspace in RTL https://bugs.webkit.org/show_bug.cgi?id=302544 rdar://problem/164740784
Reviewed by Alan Baradlay.

In RenderMathMLPadded, lspace is applied as horizontal/inline-axis space
at the “start” of the content in the direction the content flows according
to the specification. Previously, lspace was always added as a positive
horizontal shift, ignoring the element's writing direction. This caused
mpadded elements inside dir="rtl" to shift in the wrong direction.

This patch updates layoutBlock() to apply lspace relative to the inline
direction to fix the issue.

* Source/WebCore/rendering/mathml/RenderMathMLPadded.cpp:
(WebCore::RenderMathMLPadded::layoutBlock):
* LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/mpadded/mpadded-003-expected.txt: Renamed from LayoutTests/platform/glib/imported/w3c/web-platform-tests/mathml/presentation-markup/mpadded/mpadded-003-expected.txt.
* LayoutTests/platform/ios/imported/w3c/web-platform-tests/mathml/presentation-markup/mpadded/mpadded-003-expected.txt: Removed.
* LayoutTests/platform/mac/imported/w3c/web-platform-tests/mathml/presentation-markup/mpadded/mpadded-003-expected.txt: Removed.

Canonical link: https://commits.webkit.org/303070@main