[web-animations] make `KeyframeEffectStack::hasMatchingEffect()` public and use it from `Styleable`. https://bugs.webkit.org/show_bug.cgi?id=302567
Reviewed by Alan Baradlay.

We have a couple of places in `Styleable` where we iterate over effects to see if one is matching a
given condition. The `KeyframeEffectStack` class has a `hasMatchingEffect()` method to do just that,
so we make it public and use it from `Styleable` to make code a bit more expressive.

* Source/WebCore/animation/KeyframeEffectStack.h:
* Source/WebCore/style/Styleable.cpp:
(WebCore::Styleable::isRunningAcceleratedAnimationOfProperty const):
(WebCore::Styleable::isRunningAcceleratedTransformRelatedAnimation const):

Canonical link: https://commits.webkit.org/303072@main