Correct fast/multicol/vertical-rl/image-inside-nested-blocks-with-border.html https://bugs.webkit.org/show_bug.cgi?id=302606 rdar://164849032
Reviewed by Alan Baradlay.

Merge: https://chromium.googlesource.com/chromium/src.git/+/c7376a0255a90880df56c2151697a8447116452e

We have a float at the beginning of the second column. The test wanted the
float to have its left edge flush with the left edge of the multicol container,
but this is vertical-rl writing mode, so it will instead have its right edge
flush with the right edge of the multicol container, since that's where columns
start in the block progression direction. So the test "failed" and was stored
as such in the expectation file. But both the layout engine and the behavior of
getBoundingClientRect were correct all along.

Corrected the test, and removed a bunch of boilerplate and turned it into
something using js-test.js instead. In order to be consistent, this also cleans
up the vertical-lr and horiontal-tb (default) variants in the same way as well.

* LayoutTests/fast/multicol/image-inside-nested-blocks-with-border-expected.txt:
* LayoutTests/fast/multicol/image-inside-nested-blocks-with-border.html:
* LayoutTests/fast/multicol/vertical-lr/image-inside-nested-blocks-with-border-expected.txt:
* LayoutTests/fast/multicol/vertical-lr/image-inside-nested-blocks-with-border.html:
* LayoutTests/fast/multicol/vertical-rl/image-inside-nested-blocks-with-border-expected.txt:
* LayoutTests/fast/multicol/vertical-rl/image-inside-nested-blocks-with-border.html:

Canonical link: https://commits.webkit.org/303103@main