[MathML] Use font underlineThickness as MathML default rule thickness https://bugs.webkit.org/show_bug.cgi?id=302510 rdar://164693673
Reviewed by Frédéric Wang.

MathML Core defines [1] the default rule thickness as the value of the font’s
post.underlineThickness field. When this value is not available, the fallback
must be zero. WebKit previously used a legacy 0.05em heuristic for rule
thickness. This patch replaces that logic with a direct lookup of the primary
font’s underlineThickness and applies a LayoutUnit conversion. If the metric
is missing, the returned rule thickness is zero.

This aligns WebKit with MathML Core and Chromium’s behavior and removes the
legacy 0.05em fallback.

[1] https://w3c.github.io/mathml-core/#dfn-default-rule-thickness

* Source/WebCore/rendering/mathml/RenderMathMLBlockInlines.h:
(WebCore::RenderMathMLBlock::ruleThicknessFallback const):
* LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/fractions/frac-parameters-3-expected.txt: Progression
* LayoutTests/platform/glib/mathml/radical-fallback-expected.txt: Platform Specific Rebaseline (No Rendering Change)

Canonical link: https://commits.webkit.org/303108@main