[macOS] Allow multi-microphone capture rdar://163945062 https://bugs.webkit.org/show_bug.cgi?id=301870
Reviewed by Jean-Yves Avenard.

On macOS, only one VPIO unit can run at a time.
This means that we can capture with echo cancellation only one microphone at a time.
To support multi-capture, we do the following heuristic:
1. Whenever getUserMedia is called by default, echo cancellation remains on by default except if web page explicitly asks for echo cancellation to be turned on.
2. If starting to capture with echo cancellation, and VPIO is not in use, business is done as usual: the CoreAudioCaptureUnit::defaultSingleton() is used as normal.
3. If starting to capture with echo cancellation, and VPIO is in use, the CoreAudioCaptureSource(s) using the CoreAudioCaptureUnit::defaultSingleton() are migrated to a non VPIO HAL unit.
   This is being exposed to the web page by firing configurationchange event on the track.
   The new capture will start with echo cancellation on.

To implement this heurisitc, we add a willChangeCaptureDeviceTo callback from CoreAudioCaptureUnit to CoreAudioCaptureSource.
We reuse the same approach as the applyConstraints code path, except if the source is not capturing, in which case we ignore the change.

On iOS, CoreAudioCaptureUnit::defaultSingleton() is still in use and responsible to change between VPIO and RemoteIO unit.
To make this clearer, CoreAudioCaptureUnit::setEnableEchoCancellation is now only defined in IOS.

To support testing with mocks, we add a second device that supports w/o echo cancellation.

Covered by added test.

Canonical link: https://commits.webkit.org/303113@main