Implement byte stream support for fetch responses rdar://162107262 https://bugs.webkit.org/show_bug.cgi?id=300312
Reviewed by Chris Dumez.

We add support for readable byte stream as body of fetch responses and requests.
This is gated by a specific feature flag.

To implement that support, we have to change the way to read from a fetch stream body.
We remove the ReadableStreamSink interface and infrastructure (including JS-builtin), and instead use plain C++ via the ReadableStreamToSharedBufferSink class.
We change FetchBodySource as well as it is now a source for the C++ ReadableStream code path.

We add a way for a byte stream to start synchrnously (previously it would wait for the next micro task to be started), and we use a highWaterMark of 1 so that the stream will start pulling synchronously as this is the current model we have for fetch non byte streams.

Test: http/wpt/fetch/fetch-byte-stream.html and existing fetch tests.
Canonical link: https://commits.webkit.org/303115@main