[CMake] Use imported targets in the libxkbcommon find-module https://bugs.webkit.org/show_bug.cgi?id=302581
Reviewed by Fujii Hironori.

Rename the find-module for libxkbcommon as FindXkbCommon.cmake to follow
a more idiomatic naming convention, and update it to define an imported
target. Give that this is expected to be used on setups with a working
pkg-config installation, do not attempt falling back to manually finding
the library, headers, and version--as done for other such modules like
for example FindGTK.cmake.

* Source/WebKit/WPEPlatform/CMakeLists.txt:
* Source/cmake/FindXkbCommon.cmake: Renamed from Source/cmake/FindLibxkbcommon.cmake.
* Source/cmake/OptionsWPE.cmake:
* Tools/WebKitTestRunner/PlatformWPE.cmake:
* Tools/wpe/backends/PlatformWPE.cmake:

Canonical link: https://commits.webkit.org/303118@main