MediaPlayerPrivateMediaSourceAVFObjC can't be instantiated in the content process. https://bugs.webkit.org/show_bug.cgi?id=302625 rdar://164861750
Reviewed by Eric Carlson.

When MediaPlayer::load was called for MSE in the web content process,
it could only succeed if the url provided was non empty. But in the content
process, with MSE, that URL is always always empty.
So we bypass the iteration of MediaEngine and force the use of the
AVFoundationMSE engine immediately as there's only one suitable anyway:
AVFoundationMSE.
* Source/WebCore/platform/graphics/MediaPlayer.cpp:
(WebCore::MediaPlayer::load):

Canonical link: https://commits.webkit.org/303158@main