REGRESSION (280352@main): [ macOS wk2 Debug ] imported/w3c/web-platform-tests/media-source/dedicated-worker/mediasource-worker-detach-element.html is a flaky crash https://bugs.webkit.org/show_bug.cgi?id=275913 rdar://130611515
Reviewed by Eric Carlson.

When MediaSource::addSourceBuffer is called; we check if the MediaSource's state
is "open".
However, when running in a worker, there's the possibility that the MediaPlayer
has been destructed already and the message to the MediaSource's worker hasn't
yet been received.
We need to test for such case in MediaSourcePrivateAVFObjC::addSourceBuffer
and abort early as needed. The error will be handled the same as if the readyState
was "closed".

Covered by existing test imported/w3c/web-platform-tests/media-source/dedicated-worker/mediasource-worker-detach-element.html

* LayoutTests/platform/mac-wk2/TestExpectations: re-enabled test.
* Source/WebCore/Modules/mediasource/MediaSource.cpp:
(WebCore::MediaSource::createSourceBufferPrivate):
* Source/WebCore/platform/graphics/MediaSourcePrivate.h:
* Source/WebCore/platform/graphics/avfoundation/objc/MediaSourcePrivateAVFObjC.mm:
(WebCore::MediaSourcePrivateAVFObjC::addSourceBuffer):
* Source/WebKit/Shared/WebCoreArgumentCoders.serialization.in:

Canonical link: https://commits.webkit.org/303178@main