REGRESSION(302787@main): [macOS Release wk2] fast/files/blob-stream-error.html is a flaky timeout rdar://164935746 https://bugs.webkit.org/show_bug.cgi?id=302685
Reviewed by Chris Dumez.

When blob is loaded, the stream reader could be GCed.
This lead to read promises to never resolve.

To prevent this, we introduce a byte stream option to ReadableStream::createByteStream which marks the stream as reachable from opaque roots if pulling.
We do a refactoring to ReadableStream::createByteStream to pass a struct of options instead of individual parameters to improve code readability.

The ReadableStream reader (default or byob) implements a custom GC check by asking whether its ReadableStream is reachable from opaque roots.

We apply this strategy to Blob and FetchBodySource.
Covered by test no longer being flaky and existing stream tests.

We also align the blob stream source and fetch stream source by keeping the byte controller as a WeaKPtr instead of a RefPtr.
This model is better as it removes potential ref cycle (that was broken when blob loading finishes or fails previously).
The memory  model is now: ReadableStream reader -> ReadableStream -> ReadableStream controller and ReadableStream source.

Canonical link: https://commits.webkit.org/303180@main