[masonry] Add check if a style is a grid-lanes https://bugs.webkit.org/show_bug.cgi?id=302709 rdar://problem/164960515
Reviewed by Sammy Gill.

New method to check if an item is a grid-lanes display type without having to manually check if it is an inline or regular grid-lanes.

Using isDisplayFlexibleOrGridFormattingContextBox() in places, so we do not need to perform checks for each type.
These checks are getting more complex now that we have flex, grid, and grid-lands.

*  Source/WebCore/html/HTMLButtonElement.cpp
*  Source/WebCore/layout/layouttree/LayoutBox.cpp
*  Source/WebCore/layout/layouttree/LayoutBox.h
*  Source/WebCore/rendering/style/RenderStyle.h
*  Source/WebCore/rendering/style/RenderStyleInlines.h

Canonical link: https://commits.webkit.org/303225@main